Generate a crypto_hash class.
This is because we might be ending up right now with a bunch of hashes
being generated in a variety of places, and I don't want to end up
copying and pasteing stuff everywhere.

Also, this means that in the future we can end up swapping one hash
routine with another trivially.