I'm tired of unused-argument warnings. Silence a bunch of them.
The whole point of compiling -Wextra and likewise is that you find out
about these possibly incorrect portions of code, and then go and
give them the unused-argument attribute to indicate it's fine.

However it would appear the original authors haven't, so each time a
compile error occurs I have to dig through 10 warnings to discover
what went wrong. This eliminates most/all of the headers containing
unused arguments, and a variety of source files that I felt like
fiddling with.