Generate a huge map of expression ids -> functions to serialise them.
The benefit of doing this is that they get stored in CBMCs native
irep_idt string-poll type: ie, it's a number. And, in a map as well.
This means identifying what serialisation routine to use becomes
O(log n) rather than O(n). *And*, we avoid a bunch of string comparisons
from repeatedly converting from string to irep_idt .