Fix how we serialise symbol names.
If we've already seen it, use the hash of its value in this calculation.
We shouldn't be reusing its name because a) it can contain history and
b) if it doesn't, it can lead to false positives.

Also handle the case where we don't have a hash of a symbols value. This
occurs when we've made some kind of assumption or otherwise defined the
symbol without giving it a value. In that case, use its name, we have
nothing else.