Identify a piece of state by it's original symbol name:
Previously the L1 name has been used to identify a piece of state,
this is good because it has the identifier of the thread it's local to
encoded in that name, and the frame number to distinguish state when
a function is called recursively.

However, it turns out that that frame number is the number of times the
function has been invoked, not the recursion depth we're currently at.
Which means that if we have an infinite loop that calls one function
repeatedly, those frames will be all have unique frame numbers in their
state, preventing identical states being recognized as identical.
Additionally, those pieces of state never get discarded or overwritten,
and so we have useless history hashed into state.

So, use the original symbol name to store pieces of state data. This
means that when we enter a function sequentially, it doesn't include
frame numbers (or thread numbers) in the state hash, and it overwrites
the state of previous entries into the function.

This could be a lot better: performing this change means we lose all
accuracy when dealing with re-entrent or recursive functions, including
across threads.