Add routine to construct the expr for a property monitor.
This consits of taking an expression provided by ltl2ba and parsing
it, followed by replacing each symbol in there with another absolute
symbol provided by ltl2ba.

This is because we may need to generate expressions that describe both
globals and locals, and the code which parses the expression from a
string only generates symbols in a single namespace (ie, c::).