Hack ansi c parser to accept the small C snipits produced by ltl2ba:
It appears that the "namespace" class has two context objects - and
doesn't document what they do at all (woo). However, when using to_expr
on a small C snippet, the patched portion of code only takes one of
them, and it's the one that doesn't contain any symbols. So later on
typecheck_expr_symbol can't find any global variables.

We get around this by feeding the correct context to the C scanner but...
it seems that this (possibly) could lead to situations where we silently
discard symbols. Or something like that (look at how the consts work..)

Anyhoo, this needs closer investigtation at some point in time that
isn't now.