Track number of proposition changes that occur in the program.
We need to know how many times the propositions in the program have
changed - otherwise we risk missing an input state in the LTL monitor.
So, we insert a variable that increments every time a transition runs -
the monitor can now reject paths that contain more than one transition,
which would mean it's missed an input state.

To ensure this is thread safe, we place the global assignment that
triggers the proposition change, the proposition updates, and the
counter update in an atomic block, and follow it by a yield statement.
This generates the same number of context switches as before.