Fix code location ordering.
Specifically, a "sourcet" object (such as the ones used in the
execution_state struct for tracking loop unwinds) were not being ordered
properly.

If you had two sourct's a and b, a in thread 0, b in thread 1, and
the program counter in b was less than a, then both "a < b" and "b < a"
would be true. This is because when the thread_nr field in b isn't
less-than a's, we continue to the program counter check instead of
checking to see if thread_nr in each sourcet is equal or not.