Convert level2 state pointer in execution_statet to a simple object.
The objective here is to give that piece of data an actually well
defined lifetime - right now the level2 state object is /always/
leaked.

The current plan is to make the level2 state a field inside an
execution_statet object and for all related data structures to have a
reference to that field. That way, level2 state only lives as long as
an execution_statet object.