Disable "basic_symex" function body, add tripwire.
I don't know where this function is used, but it looks decidedly dodgy,
in that I can't see where its execution adds to program state (it seems
that it's always just discarded).

If the breakpoint inserted doesn't trigger and I can't find any
invocations of this function, it'll get removed.