Fix boolector build process.
The problem is this: the boolector -I directives to the compiler are
only present when building in the solvers directory. However we also
include boolector and Z3 headers from esbmc/bmc.h, and so don't get
those -I directives.

This change removes the need to include boolector.h in boolector
headers by moving code to cpp files and forward declaring some structs.