Rejig how pthread library decides whether to perform lock checking.
Previously, because every time we ran esbmc it parsed the library files
again, esbmc made some preprocessor definitions saying whether or not
to perform such checks. However as I'd like the library to be precompiled
this can't continue.

Instead we generate distinct functions that perform the work, and in
a moment will get the preprocessing stage to #define things like
pthread_mutex_lock to whichever version of pthread_mutex_lock that
we would like to be used.