Make ansi-c directory the last to be built - build changes on the way.
So, in the current revision, strcpy_strabs in the c library assigns
to a function. This is because the ansi-c frontend is supposed to be
translating that function name to special goto code (ewwwwwwwwwwww).

However that doesn't work with the goto-cc shipped from cprover.org, as
it expects the function prefix to be __CPROVER or suchlike. So we either
have to discard that functionality, hack it, or get another goto-cc
like program. As it turns out, we can build one from ESBMC as it has
a write_goto_binary routine in the utils directory.

So, the next few commits will be building such a program.