Pump mode information into mode.h header in preperation for langapi mode change:
Specifically what's going to happens is we get rid of all these "dummy"
files and fill the language mode table only with those languages that
have been actually compiled in. This reduces some of the insanity that
goes on when one wishes to link ESBMC into a smaller tool that does
something else.