Don't use dl-lib routines to locate c library symbols.
Two reasons: it seems that dlopen doesn't allow us to open our self
binary, possibly because we're an executable, not a shared object.
Other reason is that the binary object symbols get discarded,
probably because nothing links against them. They're also likely to
disappear when the binary gets stripped.

So, this way we link against those symbols and determine what clib to
use nondynamically. However this means that we can't use
add_cprover_library when there's no cprover library.