Only fetch c library symbols that are used in the program being tested.
This uses the same logic used in add_cprover_library as before I began
messing with it to determine what symbols to pull out.

Otherwise we get the whole C library pulled in, and unfortunately
this triggers ESBMC always using BV z3 logic, which causes several
tests to segfault. Ahem.