Attempt to recursively fetch symbols from C library:
If we have a C library function call a C library function, only the
first will be pulled in because it's referenced by the program under
test. Wheras we actually want any function/symbol it uses to be pulled
in too.

This change begins to try to do that, unsuccessfully.