Second attempt at making C library sym dependancies nonslow:
We generate a big list of what symbols depend on what other symbols and
store it in a multimap. Then each time we use a symbol, we remove it
from that multimap inserting the symbols it depends on into a list of
other things to insert.

That way once we've generated the list of symbol dependancies, we don't
have to iterate through ireps again.