Fix symbol dependancy resolution:
As it turns out, multimap equal_range returns non-inclusive iterators
rather than inclusive ones, so the failed-to-find result is identical
iterators.

Also, make list input to ingest_symbol a reference, we're not passing
the list around by value.

It seems performing this change (including previous commit) cuts the
runtime of a small test from 200ms to 150ms. An older version of
ESBMC takes 50ms though. IMO this is acceptable-ish overhead for the
C library behaving in the manner you might expect from a C file, rather
than pattern matching symbol names.

(The overhead in this situation really appears to be having to search
through all the ireps for symbol names. This is something we can do at
compile time (of the C library) to optimise, or an even better fix is
to fix the current IR situation to not be chronically slow.