Avoid null dereference when assigning back to return string struct.
Using string abstraction, we hand a functions that return a char pointer
another pointer to a string struct, to record information about the
string the return pointer corresponds to.

However if we don't store the return value, we pass null in instead,
in which case we need to avoid dereferencing it upon return.