Fix handling of struct **'s for Z3.
What would occur, is create_pointer_type receives a (struct blah **),
dereferences the type and creates the corresponding z3 foo. But later
on there are assumptions that there's only one level of pointing going
on.

So instead we record what the base type of the pointer is, so that we
can find it and make decisions no matter how many levels of pointers
there are.