Don't attempt to take the address of null.
Specifically, when sending the shadow return value pointer to a strabs
function, there are circumstances when the pointer it points at doesn't
exist. Specifically, if we discard a return value, then there won't be
anywhere for the return string struct ptr to be stored.

So instead of just blindly taking the address of the return variables
string struct, check if there's no return varaible, and if so, send
null instead of any particular pointer.

There are guards on instructions accessing that variable to prevent
the null from being dereferenced.