Fix irep reference count ordering.
In its current form, an irep removes its reference to some data and
decrements the reference count. Then it takes its new data and
increments the reference count to that.

All very fine and well until you get a situation such as the current
string_abstraction.cpp:427 loop, where we copy an irep, then repeatedly
delve further into it, discarding the previous record of it. This leads
to an irep decrementing the reference, freeing its own data, then
referencing data that it just freed. There's an assert to allegely
catch this when it happens, but it hasn't been firing.

So, fix the reference ordering: reference the data we'll be keeping,
then deref the data we're losing. That way there's no scope for things
being freed when we want to keep them.

My eyes.