Readjust how ansi-c directory builds.
Specifically, make all ansi-c objects and langapi/util/goto-program foo
be "order" dependancies: that is, we depend on it existing, not on the
value/file. This isn't entirely true, but the truth is that we don't want
to remake c2goto and all the c library for every single change to the
reset of esbmc.

Changes to the ansi-c dir probably should cause a rebuild, however header
changes will feed through to ansi-c files and cause a rebuild, again.

So, IMO the best way is to just build the C library; and then don't
rebuild it until you need to, upon which it's explicit. If you
accidentally need to rebuild it, you've done something wrong.