Remove all references to BV_REFINEMENT et al.
This is because it requires code that we don't have: specifically the
bv_refinement directory. It isn't in cbmc, it isn't in esbmc. All it's
going to do is get in the way (and it's impossible for me to test that
it actually works once I've converted it to the new solver form)