Use pointers to execution_statets, not values.
The regression test 12_lms_new happens to generate 2Gb of SSA steps,
which unfortunately gets duplicated when being moved between fields
in reachability_treet, doubling the memory requirements.

So, don't store by value in the big list of reachability states,
instead store a pointer to an execution state. This has the added
benefit that we reduce needless copying back and forth between
memory whenver we get an execution_statet out to run it.