Fix fix in r413, correctly collect pointer type.
It turns out that in create_pointer_type, the actual type can be taken
from the original type _or_ via any chain of subtypes, where I'd
previously assumed it was only from somewhere in the subtypes.

Fix this to set the actual_type variable in places where we've determined
what type to use.