Adjust solvers maekfile dependancies.
Make objects in dirs depend on the dir being rebuild, make the _dir_
phony. Previously the objects were phony, and each time we ran make the
solvers object would be rebuild.

This way the solvers object is only rebuild it the actual objects in a
dir are changed, and on each pass all dirs have make run in them anyway.