Further fix location-info preprocessor statements.
Seeing how cpp is using the lexer without a parser in mind, it's
difficult to hack in rules for collecting lineno information /outside/
of the lexer itself.

Another problem is that, for reasons I can't quite identify, some #if
statements are having their macros expanded at some point where if it
expands to #if 1, we recognize it as a lineno statement. Grr.

So: put lineno rules at beginning of rules looking for the '# and the
line number info itself, enter exclusive LINEINFO state when matching.
This seems to get rid of various problems, particularly the one where
the contents of #if were becoming line info.