Make it possible to store differing types of data about things that are parsed.
Specifically: the parser situation right now depends /entirely/ on irep
derived classes just being a different set of methods to apply to an
irep. This is equivalent to no typing at all.

As I move towards a situation where that's not the case (ie: operands in
an expression are stored as operands, not as another irep) we need to be
able to make the parser store exprt's and typet's seperately (ie: not as
another irep). Right now they get by by casting back and forth between
exprts and typets.

So, add a data type for holding different chunks of information about
what we've parsed. It can't be a union sadly because C++ sucks.