Eliminate parser "stack" facility.
This was storing information about what was being parsed in a vector:
parsers allocate a new item, fill it with data, then later on congeal
it or several back into a larger data structure / irep.

This is exactly the feature that yacc/bison already provide (with $$,
$1, and so forth). look_of_disapproval. Eliminate the stack facility,
I'll end up rewriting the ansi-c parser to not be so rubbish.

The current direction of irep and parser changes is going to kill the
parsers for: promela, smtlang, smvlang, cvclang, cpp, bplang. However
they're not used for anything anyway, and are just dead code adding
more maintenence overhead. So, they'll have to be disabled pretty
shortly.