Add "direct interleavings" option.
I have a problem: I'm trying to work out why the hell some piece of
isn't behaving in a certain way when two threads are interleaved. But
I can't really fetch the SSA program because I don't fully know what
interleaving it is that's the one I want.

Therefore there needs to be some mechanism where I can direct what
interleavings happen, and when. It doesn't have to be fantastically
complex or fully featured in my current situation as I can hack it, but
the target is to generate just _one_ program from a multithreaded
program, where the interleaving that I want happens.