Add show-ileave-points option.
The aim of this is to inspect all GOTO code and print the goto insns
that will cause an interleave point when executed. This is vital for
understanding how ESBMC is handling the everything.

IE: to debug some piece of code, we need to know where interleaves
might be happening so that they can either be disabled or directed
in some manner.