Don't attempt to convert "true" or "false" to int
Specifically, the values in the "value" field of boolean constants are
strings (uuuggghhh) and so binary2integer will now throw its cookies if
it's converted to a number.

The number it's converted to is not used however as it's a bool, so
there's no need to represent the value as "1" or "0".