Fixup insert_subtype const misbehavior
Sooo... in its previous instance, insert subtype was calling
find("subtype") and then assigning to the returned irep, after casting
it. Unbeknownst to me, that cast was eliminating a const. Leading to a
value being assigned to the "nil" irep, which breaks absolutely
everything else.

So in the new form, we adjust how loops iterate and don't do that.