Remove smtlang dir
Presumably this is a parser for the SMT-LIB syntax. I don't know what
purpose it serves, but it seems complete and working.

However it isn't being used by ESBMC, and it doesn't sound like it will
be any time soon (how does one model check SMT?), so off it goes. If
needs be, it can be resurrected in the future.