Fix execution_statet lifetime irregularity
IE: if we reach a terminal state, don't keep it hanging around for a
while in the _cur_target_state pointer, instead keep it temporarily
until the next state equation is requested.

A side effect of this is that we clean up the heavily coupled interface
between the esbmc bmct class and everything else.