Restructure the way execution states are explored
Specifially we continue to follow the depth-first-search algorithm that
Lucas has already implemented, but we do different parts at different
times.

We don't free any execution states once we've run to completion, we
instead pass it to the backend. Once that's done with it, we then reset
back to an unexplored state. The benefit of this is that we always then
free execution states in the reverse order to how they're allocated,
meaning that we never reset the string pool state to invalidate an
existing execution state.

Required a little beating of bmc.cpp, which is fine.