Reorganize how z3 obj gets constructed.
Specifically, pass some of the information to the constructor rather
than calling other things later, for convenience.

The ultimate aim for this branch is to only ever use one z3 context,
that we push and pop whatever we've constructed to/from. The point being
to avoid reallocating Z3 contexts repeatedly as they leak memory like
there's no tomorrow.