Hack Z3 to release memory from it's symbol table
Specifically, it has an internal global symbol table that stores
absolutely every symbol that we create with Z3_mk_string_symbol
(according to my investigation with IDA (entirely legal)).

This is particularly leaky - something in the region of 50Mb over an
hour or so, depending on the program being executed of course.

It also turns out that there's an internal API call that frees said
global symbol table - there are no internal references to in within
libz3.so. Sad face. So, call it directly instead.

We *have* to call Z3_reset_memory afterwards - the call itself doesn't
actually free anything, but it does check after its inaction that
various things are initialized, and it reinitializes the global symbol
table for us. Otherwise, some null pointers would be stomped on.

In some future world, presumably we can inform Microsof that freeing
this memory is desireable, and they can hook it into Z3_reset_memory.