Fix solvers.o dependency ordering.
For whatever reason in the previous configuration, make believed that
building the directory rule was sufficient to build solvers.o: ie, if
the directory rule rebuilt z3.o, this wasn't noticed, and was only
picked up on the next build.

So, alter this so that solvers.o always runs /all/ the directory rules,
and then starts making decision on whether the objects in those
directories have been rebuilt. Fixes the above problem.