Revert everything in dotypefixing so far
So, I've been spending a large amount of time rotating over how to best
fix the ansi-c parser type safety situation. And it's been unpleasent,
because (see the history) there's a log of undefinedness with regards to
what happens to an identifier token once it's parsed. IE, it gets shoved
into a subtype, and then later on the code to /make/ a decleration has
to go hunt for it. This is not type safe and suboptimal.

On closer inspection, I find such wonderful things such as this:

storage_class:
        TOK_TYPEDEF    { $$=$1; set($$, "typedef"); }
        blah blah blah

Now, typedef isn't a storage class. It's actively harmful to think of it
as such. And the existance of this piece of code suggests to me that the
C frontend is in fact a mishmash of collecting information together into
the internal representation, then elsewhere trying to work out what the
hell to do with it.

(Another reason to fix the internal representation. It encourages
mashing more data into an irep without thinking about what you /should/
be doing.)