Move 'declarator' NT, adjacent to its subterms
Incidentally if you're reading the commit log, the previous commit
incorrectly thought typedef wasn't a storage class and used that as
evidence the C parser is hacky to hell.

Turns out typedef *is* a storage class. On the whole though my evidence
inre the C parser is the declarator situation - it just doesn't follow
the C spec and leads to the mush of irepts being arranged.