Don't make "declarator" the subtype of, say, code.
This occurs when we have a name that's been declared, and have a
postfixing declarator that morphs what it is (say, to code, or an
array). In that case we want any type information associated with the
declarator, but we don't want the declarator itself.

So if what we _have_ is a declarator, add its subtype into the
code/array or whatever. But if we have something else (like, another
dimension of an array) then set _that_ as the subtype.