Mercilessly remove chunks of parser.
Specifically, everything related to duplicating the declarator syntax
for allowing typedef names in parameter declarations. The most offensive
thing about it is that there is no difference in what the parser does
with this data (goes straight to new_declaration), so perhaps it only
exists to clobber some harmless shift/reduce conflict. Or maybe there's
just no reason.

Anyways; now the test 01_cbmc_typedef1 fails, which is the only outcome
of deleting this chunk.