Start fixing usage of do_pointer
The problem in this circumstance was that type qualifiers were being
applied to the subtype, rather than the pointer type.

This in turn reveals the problem that do_pointer is still creating types
in reverse order (the subtype of a type becomes "pointer", rathern than
the type itself being modified, or w/e).