Handle more cases where decl irep must disappear
Specifically,

  void (*bees)(int something);

Ends up being a declarator, because it's declaring a variable with
nothing interesting about it. However,

  void *(*bees)(int something);

Becomes a pointer, because a) it's a pointer but also b) it may have
other type qualifiers associated with it.

With this, all the regression tests pass again. In the glorious
type-safe future, this will all become clearer as all sub-nterms of
identifier-declarator become associated with some kind of irep type that
has a subtype and an identifier field.

Incidentally, the history of this branch is a little embaressing, but
it's also a learning curve, so meh.