Revert work on irep branch to date:
It's a fun escapade, and I discovered some abuse in the C parser that
has now been fixed (merge incoming). However, I've also figured that
there's more drama that needs to occur to fix irep representation, due
to copy constructors.

Specifically, because the rest of CBMC/ESBMC consider all the ireps
interchangeable, they're repeatedly assigned between exprts and typets
and suchlike mercilessly. This is bad, because we'll immediately run
into copy constructor deat and other casting problems, the moment we try
to alter the shape of one irep. It's a minefield.

So a better solution for the moment is to provide accessors to data in
the ireps (such as operands(), but also value(), identifier(), blah blah
blah) and then make things like add(), remove(), find() and other war
criminals protected so that it can't be used in the rest of esbmc. Then
we can start approaching some form of typesafe situation, because we've
successfully decoupled access to data from how it's stored.

The point of this is that we get to creep in the type safety direction,
_without_ making any data representation changes that will screw stuff
up and manifest as random segmentation faults.