Finish up; couple of nonterminals I missed
So - the next step is to go through the parser looking for abuses of
terminal tokens like ':' or '*', and instead of assuming that they have
an exprt associated with them that we'll access through stack, instead
generate one immediately with newstack().

After that, we can start looking at eliminating stack() and friends.

And after that, object lifetime D: