Fix some typos (elaborate on plan)
So that plan - previous commit talked about getting rid of terminals
being used as scratchpads for expressions. I have a better idea - let
them do that, but define what the terminal represents as being a
"location" type. It's an exprt, but the significance is that further
down the line we'll only be using it for representing an expression,
which _is_ something that we want to know about.

(IE: this lets us preserve the ability of CBMC/ESBMC to collect a bunch
of location information about what it's parsed)