Make GCC asm statement rules discard data
We can't ever do anything to do with asm, so I assume that these rules
are there to allow for parsing its syntax, but not doing anything with
its contents. So, to make bisons type situation happy, add some blank
rules that end up with any parsed data being discarded when we have a
type problem.