Hack operands() type situation.
So, now I'm starting to eliminate accesses to "add", "find" and the
like, and this is one of the last cases where "operands" is accessed
directly.

However, the type here needs casting, as operands() returns a vector of
exprts, and get_sub() is a vector of irepts. Some future fixing
situation will have "components" in a similar form to expr operands, ie
we can access a vector of exprs. Until then, cast cast cast.

Also, 900 get.