Bump type() method from exprt to irept
This is so that we can use it on such things as subcomponents of a
struct, where the subtype of the components list is irept, but all the
irepts have "type" fields.

This is kind of a step backwards in that it makes getting "type"
information more abstract; but on the other hand it means we can move
all accesses to "type" to the type() method rather than add/find.
Progressive conversion, etc etc