Don't use "type" field for non-type ireps
This appears to be a portion of code that converts a goto program into
one big irep with all the relevant data stored within it. That's good
and fine. However, the type of an instruction is being stored in the
"type" field, which conflicts with the existing "type" usage.

So, use "typeid" instead.

Also, eeeewwwwwwwwwwwww