Beat ansi_c_decl class, disambiguate "value" irep
The specific problem is that ansi_c_decls use "value" to store the value
that's just been declared. And "value" is also a constants value.

So, this patch swaps out all those points where we expect to have an
ansi_c_decl and inser the correct typing/casting - then use the
decl_value method to retrieve the value of the declaration. Hopefully we
can then just not use "value" as the storage name from then onwards.