Return a reference from accessor irep methods
This is because the "get" method itself returns a reference to the
irep_idt that contains the value for the got field. If we return a
new irep_idt, which is what we're doing right now, then we can end up
with a situation where a new local variable is allocated in some method
to store the return of value() say, and then the method ends up storing
a reference to that temporary. Which is bad.

Originally I didn't want to return a reference from these methods
because I wasn't sure of the implications of accessing the data
directly. However, seeing how get(name) already does that, clearly there
can't be anything completely wrong with it.