add("size") -> accessor
This patch features a lot of tooing and frooing on account of the
tendancy to make a call along the lines of

  something(expr.add("blah"));

Where some field /might/ get initialized as it didn't exist before - or
on the other hand it might not. This is unpleasent. Instead, replace
with the creation of a tempory exprt to be operated upon, followed by
assigning that updated exprt to the relevant field of the irep.