Add some static irep identifier strings
These are the strings that are used as identifiers throughout ESBMC -
IE, anything that deals with the "id" irep method, or is used when an
irep subclass is instanciated, or so forth.

To that end they'll be seperate from any overlap with the a_ and s_
prefixed irep names.

This is the beginning of the long awaited "working out what the hell to
do about the polywhatsistssituation of all the irep situation"
situation.