Add methods to tell if irep is a particular kind
IE, we can work out whether it's a constant, a piece of code, blah blah
blah.

The current plan is to eliminate all id() and id(blah) calls; this puts
us in a situation where (embolden) we can work out what kind of scope
certain ireps are used in(endbold).

That then enables us ot work out certain other things: for example, the
ansic_declaration or whatever can only exist in the ansi-c code, stuff
like that. But more so.