Lay out class for storing RT DFS state
So for checkpoints, we want to be able to record the state that the
reachability tree is currently in, so that it can be reached again in
the future. That means recording what interleavings occur, where, and
what paths we take.

To aid sanity, we'll also record the location number of where such
interleavings occur to make everything slightly surer; but we can also
use this to print data without actually having to retrace symbolic
execution.

Restoring state will be something along the lines of reexecuting the
program and taking the interleavings described by the recorded data.