Add space to record num ileaves explored
IE: we'd like interleavings to have some kind of unique number (or: I
think it'd be useful) so reserve some space in the file for it.

Right now the interleaving number is a product of the bmc object, not
the RT, which I don't agree with, but that'll have to be hcanged at some
other point in time.