Fix an ordering problem with saving checkpoints
The previous way of getting the cur_thread field took the thread number
that was executing when the context switch occured, not the choice that
was supposed to be made. Fix this by shifting all the cur_thread numbers
back by one record.

The first cur_thread field is always 0 so it doesn't matter; the last
one is a dummy too as there are no choices from it, see comment in
patch.