Execute through new thread generation
It seems esbmc executes new threads until they make a visible
modification to state - handle this in the dfs restoration process by
continuing to run until generating_new_threads isn't set.

This requires unprivating _go_next. This is because, on the whole, the
reachability_tree and goto_symex classes relations are inverted :|