Tame an std::vector<bool> situation
Specifically, reserving and then setting a field (as done previously) in
an std::vector<bool> doesn't seem to actually perform that action - the
size is unchanged afterwards (zero).

So instead, clear the field if it exists already, or use push_back if it
doesn't. This handles both the situation where there's a new thread and
no _DFS_traversed state, and when the _DFS_travsersed state has been
restored from a checkpoint.