Don't run threads to visible insn on restore
As it turns out, the "run a thread until it executes a visible
instruction" feature that's suppoesd to occur every time you start a
thread, is dodgy: if you start more than one thread, it only runs the
most recent one.

Additionally, it seems that the state being saved when a checkpoint is
created doesn't feature such initial runs, so we should't run those runs
either.

Finally, this patch restores the DFS state after every insn. This is to
handle thread creation in the middle of a run, where it might fiddle
with _DFS_traversed.