Nuke is_goto decisions in symex_step
So according to the ICSE11 paper, we don't consider context switches
that occur in conditionals: i.e., the condition of an if statement. I'm
not quite certain why that's the case, but apparently doing this doesn't
affect the path of the program.

Whatever the case may be; the way it's implemented is broken. In the
current configuration, immediately after we execute a goto instruction
(anywhere), the next assignment will not generate a context switch. This
means that:

  if (1)
    facebees = toads;

Wouldn't generate a context switch, no matter what. Which is bad.

So, rip out this feature for now, it can be put back in in a non
breaking-everything way in the future.