Only increment cswitch counter when cswitch occurs
Previously it looks like context switch counter increments happened
whenever a context switch /could/ be generated, not when it actually
occured. So, running with --context-switch 3 would generate all the
possible interleavings that you could make with /the first three context
switch points ESBMC ran into/. _Not_ the set of interleavings you can
make with all the possible context-switch points there are

This is because we can run over a context switch point, not take it, but
still increment the counter, which means that there are fewer cswitches
permitted from this path.

Ugh.