Fix some fallout of "array" -> accessors
It seems a variety of portions of code didn't get some fixes - the parts
where we were comparing an irep id with !=, and thus converting those to
use accessors has to be done manually.

Don't know why this hasn't been picked up in the previous few commits;
perhaps my configuration on pony is different from corinth.