Un-protect a few members of basic_symex
It appears that a large lump of bmc.cpp reads symex.ns directly, which
violates the "protected" access modifier thingy. Current gcc doesn't
pick this up however, gcc34 does. Concerning.

This can be fixed by an actual restructure in the future, for the moment
just unprotect it so that compilation works.