Add --break-at option for breaking into ESBMC.
The argument forllowing --break-at is a number, corresponding to a
location number of a GOTO instruction that we are interested in. When
given on the command line, ESBMC will execute a breakpoint whenever we
symbolically execute that instruction.

Highly useful, I've been injecting different bits of debugging code in
that does exactly this, but now I figure it's significant enough of a
feature for other people to use too.