Rejuggle ART ownership in relation to BMC obj.
IE: start moving towards a situation where BMC creates an ART obj and
asks it for formulas, as opposed to the highly hacky way it's happening
now.

In this change, we allocate the ART ourselves and stash it in the symex
object, then dance around it a little. Next up, moving more
functionality into ART from symex.