Fiddle typedefs to match what TACAS tests expect
This is the problem that a number of TACAS tests have had some headers
preprocessed into them, but not others, meaning that we get a mixture of
types that ESBMC has typedef and those extracted from the system where
it was preprocessed.

Specifically, s3_clnt.blast.02.i.cil.c redefines size_t to be an
unsigned int (aka they preprocessed it on a 32 bit machine) wheras ESBMC
says unsigned long int. Horror occurs as a result.

There's also a bitfield somewhere that has a bool type within it, which
is rejected by CBMCs parser (quite rightly; you can only use int types
in bitfields). However we have to work around it, so bools are now ints