Nuke Z3 backend static's from orbit
All this gratuitous use of statics is extremely unhealthy and probably
causes various sneeking bugs; it also will cause havok in any situation
where constant propagation can't shorten an expr and one of the methods
gets called recursively.

There are a few places where it's justified (use as a private global
counter for example) but most of these are a bad plan.