Make z3 obj creating methods static
There are a variety of z3 capi functions that use Z3_mk_int_sort. It'd
be good to use the fixed object from there, which means the fixed object
has to live in the Z3_capi class. So start munging classes so that the
Z3_capi object is entirely a sub-thing of Z3_conv.