Vastly simplify create_array_type
Here, there was repeated switching as to whether or not we were using
a bitvector or an integer for the index of an array, in each part of a
big if/else/if/else statement. Which can be vastly simplified to one
check and one call to Z3_mk_array_type.