Remove some kind of special case for casted consts
The single comment related to what these clauses are for states:

  //workaround: bug in the front-end
  //to check it, run example ex10.c of NEC

However I don't know what NEC is, google reveals nothing, and removing
these blocks of code doesn't affect any regression test.

Looking at these blocks, it seems it only ever applies to where the
signedness of a constant differs, or something. Either way, I can't see
what the bug these solve could possibly be, and it's not in the
regression tests, so I'd suggest it's either gone or never existed.

And anyway, the correct fix is to fix the frontend, not hack the
backend.