Fix appending of thread PC to state hash
This is one of the awful things existing in C++ - the '+' operator has
been overloaded to concatonate strings, but you can't do anything else
with it. So the existing code which attempts to append the program
counter to the state hash doesn't work, it just messes with some
pointers.

This leads to really weird behavior, particularly in the fact that if
you compile -O3 -static then we end up with different results (for state
hashing) than if you did with out static. Ouch.